% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening_interventions.R
\name{adjust_screening_for_intervention}
\alias{adjust_screening_for_intervention}
\title{Modify Screening Matrix for A Targeted Screening Intervention}
\usage{
adjust_screening_for_intervention(e, intervention_years = 105:109,
  pop_indices, new_level)
}
\description{
Using the intervention_years as rows and the pop_indices as columns, 
select those entries in the screening matrix during the intervention 
time period and among the targeted population which are less than 
the new_level of screening and update them to the new_level.
}
