% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.pred.fun.R
\name{constructSimulationEnvironment}
\alias{constructSimulationEnvironment}
\title{Transform theta into an Environment containing Arguments Necessary for Simulation}
\usage{
constructSimulationEnvironment(theta)
}
\description{
Transform theta into an Environment containing Arguments Necessary for Simulation
}
