% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening_interventions.R
\name{extractInterventionStatistics}
\alias{extractInterventionStatistics}
\title{Extract Data from Simulation Outcomes for Screening Intervention Analysis}
\usage{
extractInterventionStatistics(e, intervention)
}
\description{
In our screening intervention analysis we will calculate the following 
values:
}
\details{
- Number of Additional Tests
- Number of Incident Cases 
- Number of Prevalent Cases
- Number Needed to Test to Avert One Additional Incident Case
- Number Needed to Test to Avert One Prevalent Case

In order to calculate these, we will extract the following data from a
simulation: prevalent_infections, incidents, tests (per year).
}
