% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening_interventions.R
\name{format_intervention_statistics}
\alias{format_intervention_statistics}
\title{Format the Intervention Statistics}
\usage{
format_intervention_statistics(df)
}
\arguments{
\item{df}{A dataframe resulting from simulate_interventions.}
}
\description{
Take the output from simulate_interventions do the following operations: 
- Calculate Additional Tests
- Calculate Change In Prevalent Infections 
- Calculate Change in
}
