% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{load_globals}
\alias{load_globals}
\title{Load Global Variables
This function loads the global variables necessary 
for the syphilis package that do not depend on the 
state (LA or MA) having been declared.}
\usage{
load_globals(model.end = 110)
}
\description{
Load Global Variables
This function loads the global variables necessary 
for the syphilis package that do not depend on the 
state (LA or MA) having been declared.
}
