% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening_interventions.R
\name{modify_simulation_environment_for_an_intervention}
\alias{modify_simulation_environment_for_an_intervention}
\title{Modify the Screening Matrix to Implement Targeted Testing and Treatment Interventions}
\usage{
modify_simulation_environment_for_an_intervention(e, intervention)
}
\description{
Check that the intervention is defined appropriately (either not at all or
as basecase, or as one of the interventions in the interventions$codename
column). Then use a switch on the intervention to call the
adjust_screening_for_intervention function with the right arguments.
}
