% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.pred.fun.R
\name{prediction.epi}
\alias{prediction.epi}
\title{Run a Simulation with Parameters theta and the DiffEq Model in syph_trans_model.cpp}
\usage{
prediction.epi(theta)
}
\arguments{
\item{theta}{A named numeric vector defining each of the model parameters.}
}
\description{
This function uses three steps: 
1. constructSimulationEnvironment: This produces an environment object containing the 
   necessary arguments for the Rcpp function syphSim defined in syph_trans_model.cpp
2. runSimulation: Using the objects in the environment `e`, run syphSim.
3. summarizeSimulation: Run model.prev to extract summary features from the simulation.
}
\details{
This approach is preferred to a less modular (i.e. combining all the above
steps into one function) approach because in between steps 1 and 2 the simulation 
environment can be edited (for example to define intervention scenarios, to change or define 
debugging arguments, etc.), and if alternative summary measures are desired step 3 can be 
changed to any other function that analyzes the output from syphSim.
}
