% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening_interventions.R
\name{simulate_interventions}
\alias{simulate_interventions}
\title{Simulate Each Intervention with a Given Parametrization}
\usage{
simulate_interventions(theta,
  interventions_list = c("msm_annual_hr_msm_quarterly",
  "prior_diagnosis_quarterly", "high_activity_quarterly"), info_extractor)
}
\description{
Go through each of the interventions (including the basecase),
simulate them, extract the Intervention Statistics, and 
construct a dataframe with the results.
}
