% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syphilis.R
\docType{package}
\name{syphilis-package}
\alias{syphilis}
\alias{syphilis-package}
\title{A model of syphilis fit to data from Louisiana and Massachusetts}
\description{
A model of syphilis fit to data from Louisiana and Massachusetts
}
\author{
\strong{Maintainer}: Christian Testa \email{ctesta@hsph.harvard.edu}

Authors:
\itemize{
  \item Jessie Lowell
  \item Ashleigh Tuite
}

}
