% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syphLAMA.R
\docType{package}
\name{syphLAMA-package}
\alias{syphLAMA}
\alias{syphLAMA-package}
\title{syphLAMA A model of syphilis fit to data from Louisiana and Massachusetts}
\description{
syphLAMA A model of syphilis fit to data from Louisiana and Massachusetts
}
\author{
\strong{Maintainer}: Christian Testa \email{ctesta@hsph.harvard.edu}

Authors:
\itemize{
  \item Jessie Lowell
  \item Ashleigh Tuite
}

}
